Move `AsAny` to `util`
