+++ src/network.rs
@@ -1,4 +1,4 @@
-use crate::node::*;
+use crate::node::{AnyNode, StartNode, Node, NodeLike};
 use crate::util::error_f;
 
 pub type LayerT = Vec<AnyNode>;
